requirements.txt
Django==1.10.2
djangorestframework==3.5.3
psycopg2==2.6.2
requests==2.12.4
